# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fanz",
        "version": "1.0.0"
    },
    "paths": {
        "/api/players": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get a list of players",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Create a new player",
                "operationId": "4e2752a0305de0adfa2a8c9ac1aa37a9",
                "requestBody": {
                    "description": "Player data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "team_id": {
                                        "type": "integer"
                                    },
                                    "position_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/players/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get a player by ID",
                "operationId": "dbef515d1b5f749e736bca18efe444e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "api",
                "scheme": "bearer"
            }
        }
    }
}